API: POST improved - patients are inserted including risk factors!

+++ src/MedicalRecords.Domain/Services/PatientService.cs
@@ -48,12 +48,26 @@ namespace MedicalRecords.Domain.Services
 
         public async Task<PatientResponse> AddPatientAsync(AddPatientRequest request)
         {
-
             var patient = _patientMapper.Map(request);
 
-            var result = _patientRepository.Add(patient);
+            Patient result = _patientRepository.Add(patient);
 
             await _patientRepository.UnitOfWork.SaveChangesAsync();
+
+
+            // Create new records in joining table:
+            if (request.RiskFactors != null && result != null)
+            {
+                foreach (GetRiskFactorRequest riskFactor in request.RiskFactors)
+                {
+                    AddPatientRiskFactorRequest addPRFRequest = new AddPatientRiskFactorRequest
+                    {
+                        PatientId = result.Id,
+                        RiskFactorId = riskFactor.Id
+                    };
+                    await _patientRiskFactorService.AddPatientRiskFactor(addPRFRequest);
+                }
+            }
             return _patientMapper.Map(result);
         }
 
@@ -70,15 +84,19 @@ namespace MedicalRecords.Domain.Services
             var result = _patientRepository.Update(entity);
 
             // Delete old records in joining table for existing record:
-            foreach (PatientRiskFactor patientRiskFactor in existingRecord.PatientRiskFactors)
+            if (existingRecord.PatientRiskFactors != null)
             {
-                DeletePatientRiskFactorRequest deletePRF = new DeletePatientRiskFactorRequest
+                foreach (PatientRiskFactor patientRiskFactor in existingRecord.PatientRiskFactors)
                 {
-                    PatientId = patientRiskFactor.PatientId,
-                    RiskFactorId = patientRiskFactor.RiskFactorId
-                };
-                await _patientRiskFactorService.DeletePatientRiskFactor(deletePRF);
+                    DeletePatientRiskFactorRequest deletePRF = new DeletePatientRiskFactorRequest
+                    {
+                        PatientId = patientRiskFactor.PatientId,
+                        RiskFactorId = patientRiskFactor.RiskFactorId
+                    };
+                    await _patientRiskFactorService.DeletePatientRiskFactor(deletePRF);
+                }
             }
+            
 
             // Create new records in joining table:
             if (request.RiskFactors != null)
@@ -92,9 +110,6 @@ namespace MedicalRecords.Domain.Services
                     };
                     await _patientRiskFactorService.AddPatientRiskFactor(addPRFRequest);
                 }
-            }else
-            {
-                Console.WriteLine("PPPPPPPP");
             }            
 
             // Confirm changes:
